moving code around and updating cmake lists file to support build_simple.sh and Mac OS X building.
