+++ src/CMakeLists.txt
@@ -191,7 +191,6 @@ list(APPEND SRCS
     # ------------------------------------------------------
     limit/challenge.cc
     limit/rl_obj.cc
-    limit/rl_op.cc
     limit/enforcer.cc
     limit/config.cc
     limit/configs.cc
