--- src/limit/rl_op.cc
@@ -1,372 +1,0 @@
-//: ----------------------------------------------------------------------------
-//: Copyright (C) 2016 Verizon.  All Rights Reserved.
-//: All Rights Reserved
-//:
-//: \file:    op.cc
-//: \details: TODO
-//: \author:  Reed P. Morrison
-//: \date:    12/03/2018
-//:
-//:   Licensed under the Apache License, Version 2.0 (the "License");
-//:   you may not use this file except in compliance with the License.
-//:   You may obtain a copy of the License at
-//:
-//:       http://www.apache.org/licenses/LICENSE-2.0
-//:
-//:   Unless required by applicable law or agreed to in writing, software
-//:   distributed under the License is distributed on an "AS IS" BASIS,
-//:   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-//:   See the License for the specific language governing permissions and
-//:   limitations under the License.
-//:
-//: ----------------------------------------------------------------------------
-//: ----------------------------------------------------------------------------
-//: includes
-//: ----------------------------------------------------------------------------
-#include "limit.pb.h"
-#include "waflz/limit/rl_obj.h"
-#include "op/regex.h"
-#include "op/nms.h"
-#include "limit/rl_op.h"
-#include <fnmatch.h>
-namespace ns_waflz
-{
-//: ----------------------------------------------------------------------------
-//: \details  run a limit operator on some data
-//: \l_retval number of entries added to ao_match_list
-//:           -1 on failure
-//: \param    TODO
-//: ----------------------------------------------------------------------------
-int32_t rl_run_op(bool &ao_matched,
-                  const waflz_pb::op_t &a_op,
-                  const char *a_data,
-                  uint32_t a_len,
-                  bool a_case_insensitive)
-{
-        // assume operator is STREQ
-        ao_matched = false;
-        waflz_pb::op_t_type_t l_op_type = waflz_pb::op_t_type_t_STREQ;
-        if(a_op.has_type())
-        {
-                // operator type actually provided
-                l_op_type = a_op.type();
-        }
-        switch (l_op_type)
-        {
-        // -------------------------------------------------
-        // RX (regex)
-        // -------------------------------------------------
-        case waflz_pb::op_t_type_t_RX:
-        {
-                // -----------------------------------------
-                // get regex
-                // -----------------------------------------
-                if(!a_op.has__reserved_1())
-                {
-                        return WAFLZ_STATUS_ERROR;
-                }
-                regex *l_rx = (regex *)(a_op._reserved_1());
-                if(!l_rx)
-                {
-                        return WAFLZ_STATUS_ERROR;
-                }
-                // -----------------------------------------
-                // match?
-                // -----------------------------------------
-                //TRC_ALL("RX[%p]: %s == %.*s\n", l_rx, l_rx->get_regex_string().c_str(), (int)a_len, a_data);
-                int l_s;
-                l_s = l_rx->compare(a_data, a_len);
-                // if failed to match
-                if(l_s < 0)
-                {
-                        break;
-                }
-                ao_matched = true;
-                break;
-        }
-        // -------------------------------------------------
-        // STREQ
-        // -------------------------------------------------
-        case waflz_pb::op_t_type_t_STREQ:
-        {
-                const std::string &l_op_match = a_op.value();
-                uint32_t l_len = l_op_match.length();
-                if(l_len != a_len)
-                {
-                        break;
-                }
-                int l_cmp = 0;
-                if(a_case_insensitive)
-                {
-                        l_cmp = strncasecmp(l_op_match.c_str(), a_data, l_len);
-                }
-                else
-                {
-                        l_cmp = strncmp(l_op_match.c_str(), a_data, l_len);
-                }
-                if(l_cmp == 0)
-                {
-                        // matched
-                        ao_matched = true;
-                        break;
-                }
-                //TRACE("Got data: '%.*s' and match '%s'", SUBBUF_FORMAT(a_data), l_op_match.c_str());
-                break;
-        }
-        // -------------------------------------------------
-        // PM
-        // -------------------------------------------------
-        case waflz_pb::op_t_type_t_PM:
-        {
-                // -----------------------------------------
-                // substring match multiple strings
-                // -----------------------------------------
-                // most naive possible implementation
-                // TODO explore using pm operator
-                // -----------------------------------------
-                for(int i_val = 0; i_val < a_op.values_size(); ++i_val)
-                {
-                        // for each value
-                        const std::string& l_op_match = a_op.values(i_val);
-                        if(l_op_match.length() > a_len)
-                        {
-                                continue;
-                        }
-                        // If match value empty, continue
-                        if(l_op_match.empty())
-                        {
-                                continue;
-                        }
-                        const char *l_match = 0;
-                        if(a_case_insensitive)
-                        {
-                                l_match = strcasestr(a_data, l_op_match.c_str());
-                        }
-                        else
-                        {
-                                l_match = strstr(a_data, l_op_match.c_str());
-                        }
-                        if(l_match != NULL)
-                        {
-                                // matched
-                                ao_matched = true;
-                                break;
-                        }
-                        //TRACE("Got data: '%.*s' and match[%u] '%s'", SUBBUF_FORMAT(a_data), i_val, l_op_match.c_str());
-                }
-                break;
-        }
-        // -------------------------------------------------
-        // GLOB (glob -wildcard match)
-        // -------------------------------------------------
-        case waflz_pb::op_t_type_t_GLOB:
-        {
-                int l_flags = FNM_NOESCAPE;
-                if(a_case_insensitive)
-                {
-                        l_flags |= FNM_CASEFOLD;
-                }
-                int l_cmp;
-                const std::string &l_op_match = a_op.value();
-                l_cmp = fnmatch(l_op_match.c_str(), a_data, l_flags);
-                if(l_cmp == 0)
-                {
-                        // matched
-                        ao_matched = true;
-                }
-                break;
-        }
-        // -------------------------------------------------
-        // TODO
-        // -------------------------------------------------
-        case waflz_pb::op_t_type_t_IPMATCH:
-        {
-                // -----------------------------------------
-                // get regex
-                // -----------------------------------------
-                if(!a_op.has__reserved_1())
-                {
-                        return WAFLZ_STATUS_ERROR;
-                }
-                nms *l_nms = (nms *)(a_op._reserved_1());
-                if(!l_nms)
-                {
-                        return WAFLZ_STATUS_ERROR;
-                }
-                // -----------------------------------------
-                // match?
-                // -----------------------------------------
-                int32_t l_s;
-                l_s = l_nms->contains(ao_matched, a_data, a_len);
-                // if failed to match
-                if(l_s < 0)
-                {
-                        break;
-                }
-                break;
-        }
-        // -------------------------------------------------
-        // Exact Match list (EM)
-        // -------------------------------------------------
-        case waflz_pb::op_t_type_t_EM:
-        {
-                // -----------------------------------------
-                // get str set
-                // -----------------------------------------
-                if(!a_op.has__reserved_1())
-                {
-                        return WAFLZ_STATUS_ERROR;
-                }
-                // -----------------------------------------
-                // case insensitive
-                // -----------------------------------------
-                if(a_op.is_case_insensitive())
-                {
-                        data_case_i_set_t *l_ds = (data_case_i_set_t *)(a_op._reserved_1());
-                        if(!l_ds)
-                        {
-                                return WAFLZ_STATUS_ERROR;
-                        }
-                        // ---------------------------------
-                        // match?
-                        // ---------------------------------
-                        data_t l_d;
-                        l_d.m_data = a_data;
-                        l_d.m_len = a_len;
-                        data_set_t::const_iterator i_d = l_ds->find(l_d);
-                        if((i_d != l_ds->end()) &&
-                           (i_d->m_len == l_d.m_len))
-                        {
-                                ao_matched = true;
-                        }
-                }
-                // -----------------------------------------
-                // case sensitive
-                // -----------------------------------------
-                else
-                {
-                        data_set_t *l_ds = (data_set_t *)(a_op._reserved_1());
-                        if(!l_ds)
-                        {
-                                return WAFLZ_STATUS_ERROR;
-                        }
-                        // ---------------------------------
-                        // match?
-                        // ---------------------------------
-                        data_t l_d;
-                        l_d.m_data = a_data;
-                        l_d.m_len = a_len;
-                        data_set_t::const_iterator i_d = l_ds->find(l_d);
-                        if((i_d != l_ds->end()) &&
-                           (i_d->m_len == l_d.m_len))
-                        {
-                                ao_matched = true;
-                        }
-                }
-                break;
-        }
-        // -------------------------------------------------
-        // default
-        // -------------------------------------------------
-        default:
-        {
-                // do nothing...
-                return WAFLZ_STATUS_OK;
-        }
-        }
-        if(a_op.is_negated())
-        {
-                // negate value
-                ao_matched = !ao_matched;
-        }
-        // -------------------------------------------------
-        // TODO -push matches???
-        // -------------------------------------------------
-        return WAFLZ_STATUS_OK;
-}
-//: ----------------------------------------------------------------------------
-//: \details check if request "in scope"
-//: \return  true if in scope
-//:          false if not in scope
-//: \param   a_scope TODO
-//: \param   a_ctx   TODO
-//: ----------------------------------------------------------------------------
-int32_t in_scope(bool &ao_match,
-                 const waflz_pb::scope &a_scope,
-                 rqst_ctx *a_ctx)
-{
-        ao_match = false;
-        if(!a_ctx)
-        {
-                return WAFLZ_STATUS_ERROR;
-        }
-        // -------------------------------------------------
-        // host
-        // -------------------------------------------------
-        if(a_scope.has_host() &&
-           a_scope.host().has_type() &&
-           (a_scope.host().has_value() ||
-            a_scope.host().values_size()))
-        {
-                const data_t &l_d = a_ctx->m_host;
-                if(!l_d.m_data ||
-                   !l_d.m_len)
-                {
-                        return WAFLZ_STATUS_OK;
-                }
-                bool l_matched = false;
-                int32_t l_s;
-                l_s = rl_run_op(l_matched,
-                                a_scope.host(),
-                                l_d.m_data,
-                                l_d.m_len,
-                                true);
-                if(l_s != WAFLZ_STATUS_OK)
-                {
-                        return WAFLZ_STATUS_ERROR;
-                }
-                if(!l_matched)
-                {
-                        return WAFLZ_STATUS_OK;
-                }
-        }
-        // -------------------------------------------------
-        // path
-        // -------------------------------------------------
-        if(a_scope.has_path() &&
-           a_scope.path().has_type() &&
-           (a_scope.path().has_value() ||
-            a_scope.path().values_size()))
-        {
-                data_t l_d = a_ctx->m_uri;
-                if(!l_d.m_data ||
-                   !l_d.m_len)
-                {
-                        return WAFLZ_STATUS_OK;
-                }
-                // use length w/o q string
-                // use length w/o q string
-                if(a_ctx->m_uri_path_len)
-                {
-                        l_d.m_len = a_ctx->m_uri_path_len;
-                }
-                bool l_matched = false;
-                int32_t l_s;
-                l_s = rl_run_op(l_matched,
-                                a_scope.path(),
-                                l_d.m_data,
-                                l_d.m_len,
-                                true);
-                if(l_s != WAFLZ_STATUS_OK)
-                {
-                        return WAFLZ_STATUS_ERROR;
-                }
-                if(!l_matched)
-                {
-                        return WAFLZ_STATUS_OK;
-                }
-        }
-        ao_match = true;
-        return WAFLZ_STATUS_OK;
-}
-}
